+++ src/FSharp.Analyzers.Cli/SarifReport.fs
@@ -1,0 +1,106 @@
+module FSharp.Analyzers.Cli.SarifReport
+
+open System
+open System.IO
+open Microsoft.CodeAnalysis.Sarif
+open Microsoft.CodeAnalysis.Sarif.Writers
+open Microsoft.Extensions.Logging
+open FSharp.Analyzers.SDK
+
+let writeReport (logger: ILogger) (results: AnalyzerMessage list) (codeRoot: Uri) (report: string) =
+    try
+        // Construct full path to ensure path separators are normalized.
+        let report = Path.GetFullPath report
+        // Ensure the parent directory exists
+        let reportFile = FileInfo(report)
+        reportFile.Directory.Create()
+
+        let driver = ToolComponent()
+        driver.Name <- "Ionide.Analyzers.Cli"
+        driver.InformationUri <- Uri("https://ionide.io/FSharp.Analyzers.SDK/")
+
+        driver.Version <-
+            string<Version>(System.Reflection.Assembly.GetExecutingAssembly().GetName().Version)
+
+        let tool = Tool()
+        tool.Driver <- driver
+        let run = Run()
+        run.Tool <- tool
+
+        use sarifLogger =
+            new SarifLogger(
+                report,
+                logFilePersistenceOptions =
+                    (FilePersistenceOptions.PrettyPrint
+                     ||| FilePersistenceOptions.ForceOverwrite),
+                run = run,
+                levels = BaseLogger.ErrorWarningNote,
+                kinds = BaseLogger.Fail,
+                closeWriterOnDispose = true
+            )
+
+        sarifLogger.AnalysisStarted()
+
+        for analyzerResult in results do
+            let reportDescriptor = ReportingDescriptor()
+            reportDescriptor.Id <- analyzerResult.Message.Code
+            reportDescriptor.Name <- analyzerResult.Message.Message
+
+            analyzerResult.ShortDescription
+            |> Option.iter (fun shortDescription ->
+                reportDescriptor.ShortDescription <-
+                    MultiformatMessageString(shortDescription, shortDescription, dict [])
+            )
+
+            analyzerResult.HelpUri
+            |> Option.iter (fun helpUri -> reportDescriptor.HelpUri <- Uri(helpUri))
+
+            let result = Result()
+            result.RuleId <- reportDescriptor.Id
+
+            result.Level <-
+                match analyzerResult.Message.Severity with
+                | Severity.Info -> FailureLevel.Note
+                | Severity.Hint -> FailureLevel.Note
+                | Severity.Warning -> FailureLevel.Warning
+                | Severity.Error -> FailureLevel.Error
+
+            let msg = Message()
+            msg.Text <- analyzerResult.Message.Message
+            result.Message <- msg
+
+            let physicalLocation = PhysicalLocation()
+
+            physicalLocation.ArtifactLocation <-
+                let al = ArtifactLocation()
+                al.Uri <- codeRoot.MakeRelativeUri(Uri(analyzerResult.Message.Range.FileName))
+                al
+
+            physicalLocation.Region <-
+                let r = Region()
+                r.StartLine <- analyzerResult.Message.Range.StartLine
+
+                r.StartColumn <-
+                    analyzerResult.Message.Range.StartColumn
+                    + 1
+
+                r.EndLine <- analyzerResult.Message.Range.EndLine
+
+                r.EndColumn <-
+                    analyzerResult.Message.Range.EndColumn
+                    + 1
+
+                r
+
+            let location: Location = Location()
+            location.PhysicalLocation <- physicalLocation
+            result.Locations <- [| location |]
+
+            sarifLogger.Log(reportDescriptor, result, System.Nullable())
+
+        sarifLogger.AnalysisStopped(RuntimeConditions.None)
+
+        sarifLogger.Dispose()
+    with ex ->
+        logger.LogError(ex, "Could not write sarif to {report}", report)
+        logger.LogInformation("{0}", ex)